requirements_manage.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioredis==1.3.1
aiosignal==1.3.1
asgiref==3.4.1
async-timeout==4.0.3
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
beautifulsoup4==4.9.3
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
charset-normalizer==3.3.2
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cycler==0.11.0
daphne==3.0.2
dash==1.21.0
dash-bootstrap-components==0.13.0
dash_core_components==1.17.1
dash_html_components==1.1.4
dash_renderer==1.9.1
dash_table==4.12.0
Django==3.2.6
django-bootstrap4==3.0.1
django-plotly-dash==1.6.4
django-redis==5.0.0
dpd-static-support==0.0.5
dpd_components==0.1.0
Flask==2.0.1
Flask-Compress==1.10.1
frozenlist==1.4.1
future==0.18.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==2.1.1
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.1.5
matplotlib==3.3.4
msgpack==1.0.2
multidict==6.0.5
mysql-connector-python==8.0.29
mysqlclient==2.0.3
numpy==1.19.5
openai==0.28.0
pandas==1.1.2
Pillow==8.4.0
plotly==5.2.2
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2021.1
redis==3.5.3
requests==2.32.3
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tenacity==8.0.1
tqdm==4.66.5
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==2.0.3
yarl==1.9.4
zipp==3.20.0
zope.interface==7.0.1
